fix EditBeneficiaryModal Validation
